refactor(*): sections 폴더 추가, try/catch로 예외 처리
- sections: 각 메뉴의 섹션들을 분리
- try/catch로 error를 받아 메시지 출력
