+++ src/components/VendingMachineManageMenu/sections/ChargedCoinSection.js
