+++ src/configs/constants.js
@@ -31,4 +31,5 @@ export const EXCEPTIONS = {
   OUT_OF_STOCK: new Error('재고가 없습니다.'),
   NO_SUCH_ITEM: new Error('아이템 조회에 실패했습니다.'),
   NOT_ENOUGH_MONEY: new Error('금액이 부족합니다.'),
+  WRONG_CHARGE_AMOUNT: new Error('올바른 금액을 투입해주세요.'),
 };
